# dist/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cities": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/flights/checkin": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    },
    "/airports/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cityname",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/airports/{code}/inbound": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    },
    "/airports/{code}/outbound": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    },
    "/routes/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "point_a",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "point_b",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departure_date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "booking_class",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "connections",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    },
    "/routes/booking": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    }
  }
}